requirements.txt
Flask== 1.1.2
gunicorn==19.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
WerkZeug==1.0.1
numpy>=1.19.2
pandas>=1.1.3
keras-facenet==0.3.2
keras==2.4.3
opencv-python==4.4.0.46
pillow==8.0.1
mtcnn==0.1.0
requests==2.24.0

